fix: some final fixes for modals
