+++ src/components/BurgerIngredients/BurgerIngredients.js
@@ -11,7 +11,7 @@ import style from "./BurgerIngredients.module.css";
 
 import { Tab } from "@ya.praktikum/react-developer-burger-ui-components";
 
-import { handleItemSearch } from "../../utils/findItem";
+import { handleItemSearchWithId } from "../../utils/findItem";
 
 import IngredientCard from "./IngredientCard/IngredientCard";
 
@@ -22,7 +22,7 @@ function BurgerIngredients() {
   const dispatch = useDispatch();
 
   const manageIngredientModal = (isOpen) =>
-    dispatch(handleIngredientModal(isOpen));
+    dispatch(handleIngredientModal({ isOpen }));
   const manageIngredient = (ingredient) =>
     dispatch(handleCurrentIngredient({ ingredient }));
 
@@ -49,12 +49,8 @@ function BurgerIngredients() {
     setCurrentTab(closestTitle);
   }
 
-  function handleIngredientClick(e) {
-    const parentElement = e.target.parentElement.querySelector(
-      ".text_type_main-default"
-    );
-    const target = parentElement && parentElement.textContent;
-    const item = handleItemSearch(data, target);
+  function handleIngredientClick(id) {
+    const item = handleItemSearchWithId(data, id);
 
     if (item) {
       manageIngredientModal(true);
@@ -114,9 +110,9 @@ function BurgerIngredients() {
                   (el, index) =>
                     el.type === "bun" && (
                       <Link
-                        className={style.card}
-                        onClick={handleIngredientClick}
                         key={el._id}
+                        onClick={() => handleIngredientClick(el._id)}
+                        className={style.card}
                         to={{
                           pathname: `/ingredients/${el._id}`,
                           state: { background: location },
@@ -141,9 +137,9 @@ function BurgerIngredients() {
                   (el, index) =>
                     el.type === "sauce" && (
                       <Link
-                        className={style.card}
-                        onClick={handleIngredientClick}
                         key={el._id}
+                        onClick={() => handleIngredientClick(el._id)}
+                        className={style.card}
                         to={{
                           pathname: `/ingredients/${el._id}`,
                           state: { background: location },
@@ -172,9 +168,9 @@ function BurgerIngredients() {
                   (el, index) =>
                     el.type === "main" && (
                       <Link
-                        className={style.card}
-                        onClick={handleIngredientClick}
                         key={el._id}
+                        onClick={() => handleIngredientClick(el._id)}
+                        className={style.card}
                         to={{
                           pathname: `/ingredients/${el._id}`,
                           state: { background: location },
