+++ src/components/Modal/Modal.module.css
@@ -14,8 +14,8 @@
   top: 50%;
   left: 50%;
   transform: translate(-50%, -50%);
-  background-color: #1c1c21;
-  z-index: 2;
+  background-color: hsl(240, 8%, 12%);
+  z-index: 20;
   width: 720px;
   min-height: 539px;
   border: 1px solid rgba(76, 76, 255, 0.2);
