+++ src/components/Modal/ModalOverlay/ModalOverlay.module.css
@@ -5,5 +5,5 @@
   bottom: 0;
   left: 0;
   right: 0;
-  z-index: 2;
+  z-index: 10;
 }
