+++ src/components/OrdersList/OrdersList.js
@@ -1,15 +1,8 @@
 import React from "react";
-import { useDispatch } from "react-redux";
 import PropTypes from "prop-types";
 import { Link, useLocation } from "react-router-dom";
 
 import OrderCard from "./OrderCard/OrderCard";
-
-import { handleItemSearchWithId } from "../../utils/findItem";
-import {
-  handleCurrentOrder,
-  handleOrderCardModal,
-} from "../../services/reducers/orderCardModal";
 
 import data from "../../assets/orders-mock-data.json";
 
@@ -17,21 +10,6 @@ import style from "./OrdersList.module.css";
 
 function OrdersList({ isProfile, onCardClick, path }) {
   const location = useLocation();
-  const dispatch = useDispatch();
-
-  const manageOrderCardModal = (isOpen) =>
-    dispatch(handleOrderCardModal({ isOpen }));
-  const manageCurrentOrder = (order) => dispatch(handleCurrentOrder({ order }));
-
-  function handleOrderCardClick(id) {
-    const item = handleItemSearchWithId(data, id);
-    console.log(item);
-
-    if (item) {
-      manageOrderCardModal(true);
-      manageCurrentOrder(item);
-    }
-  }
 
   return (
     <section className={`${style.section} ${isProfile && style.section_big}`}>
@@ -44,7 +22,7 @@ function OrdersList({ isProfile, onCardClick, path }) {
         {data.map((el) => (
           <Link
             key={el._id}
-            onClick={() => handleOrderCardClick(el._id)}
+            onClick={() => onCardClick(el._id)}
             className={style.card}
             to={{
               pathname: `${path}${el._id}`,
